+++ src/index.test.js
@@ -39,7 +39,7 @@ describe('jss-isolate', () => {
           '30%': {top: 30},
           '60%, 70%': {top: 80}
         },
-        '@supports ( display: flexbox )': {}
+        '@supports ( display: flexbox )': {},
       })
     })
 
@@ -51,7 +51,7 @@ describe('jss-isolate', () => {
   describe('works with classes', () => {
     let sheet
 
-    beforeEach((done) => {
+    beforeEach(() => {
       sheet = jss.createStyleSheet({
         link: {
           color: 'red',
@@ -60,7 +60,6 @@ describe('jss-isolate', () => {
           color: 'blue'
         }
       })
-      setTimeout(done)
     })
 
     it('should add selectors to the reset rule', () => {
@@ -80,7 +79,7 @@ describe('jss-isolate', () => {
     let sheet1
     let sheet2
 
-    beforeEach((done) => {
+    beforeEach(() => {
       sheet1 = jss.createStyleSheet({
         link: {
           color: 'red'
@@ -91,7 +90,6 @@ describe('jss-isolate', () => {
           color: 'blue'
         }
       })
-      setTimeout(done)
     })
 
     it('should add selectors to the reset rule', () => {
@@ -102,7 +100,7 @@ describe('jss-isolate', () => {
   })
 
   describe('global option "isolate"', () => {
-    beforeEach((done) => {
+    beforeEach(() => {
       jss = create().use(isolate({
         isolate: false
       }))
@@ -112,7 +110,6 @@ describe('jss-isolate', () => {
           color: 'blue'
         }
       })
-      setTimeout(done)
     })
 
     it('should use global option', () => {
@@ -124,7 +121,7 @@ describe('jss-isolate', () => {
     let sheet1
     let sheet2
 
-    beforeEach((done) => {
+    beforeEach(() => {
       sheet1 = jss.createStyleSheet({
         link: {
           color: 'red'
@@ -135,7 +132,6 @@ describe('jss-isolate', () => {
           color: 'blue'
         }
       }, {isolate: false})
-      setTimeout(done)
     })
 
     it('should not add selectors to the reset rule', () => {
@@ -148,14 +144,13 @@ describe('jss-isolate', () => {
   describe('isolate rules if they have isolate: true even if StyleSheet options is isolate: false', () => {
     let sheet
 
-    beforeEach((done) => {
+    beforeEach(() => {
       sheet = jss.createStyleSheet({
         link: {
           isolate: true,
           color: 'blue'
         }
       }, {isolate: false})
-      setTimeout(done)
     })
 
     it('should add selectors to the reset rule', () => {
@@ -167,7 +162,7 @@ describe('jss-isolate', () => {
   describe('isolate option as a string', () => {
     let sheet
 
-    beforeEach((done) => {
+    beforeEach(() => {
       sheet = jss.createStyleSheet({
         root: {
           color: 'blue'
@@ -176,7 +171,6 @@ describe('jss-isolate', () => {
           color: 'red'
         }
       }, {isolate: 'root'})
-      setTimeout(done)
     })
 
     it('should only isolate rules with matching name', () => {
@@ -189,7 +183,7 @@ describe('jss-isolate', () => {
   describe('ignore rules if property isolate is set to false', () => {
     let sheet
 
-    beforeEach((done) => {
+    beforeEach(() => {
       sheet = jss.createStyleSheet({
         link: {
           color: 'red'
@@ -199,7 +193,6 @@ describe('jss-isolate', () => {
           isolate: false
         }
       })
-      setTimeout(done)
     })
 
     it('should add selectors to the reset rule', () => {
@@ -216,7 +209,7 @@ describe('jss-isolate', () => {
   describe('don\'t duplicate selectors', () => {
     let sheet
 
-    beforeEach((done) => {
+    beforeEach(() => {
       sheet = jss.createStyleSheet({
         link: {
           color: 'blue'
@@ -227,7 +220,6 @@ describe('jss-isolate', () => {
           }
         }
       })
-      setTimeout(done)
     })
 
     it('should add selectors to the reset rule', () => {
@@ -237,7 +229,7 @@ describe('jss-isolate', () => {
   })
 
   describe('option "reset={width}" with custom props', () => {
-    beforeEach((done) => {
+    beforeEach(() => {
       jss = create().use(isolate({
         reset: {
           width: '1px'
@@ -249,7 +241,6 @@ describe('jss-isolate', () => {
           color: 'blue'
         }
       })
-      setTimeout(done)
     })
 
     it('should add width prop to the reset rule', () => {
@@ -279,7 +270,7 @@ describe('jss-isolate', () => {
   describe('nested media queries with jss-nested', () => {
     let sheet
 
-    beforeEach((done) => {
+    beforeEach(() => {
       jss = create().use(isolate(), nested())
       sheet = jss.createStyleSheet({
         link: {
@@ -289,7 +280,6 @@ describe('jss-isolate', () => {
           }
         }
       })
-      setTimeout(done)
     })
 
     it('should add selectors to the reset rule', () => {
@@ -299,7 +289,7 @@ describe('jss-isolate', () => {
   })
 
   describe('nested media queries with jss-nested with isolate:false', () => {
-    beforeEach((done) => {
+    beforeEach(() => {
       jss = create().use(isolate(), nested())
       jss.createStyleSheet({
         link: {
@@ -310,7 +300,6 @@ describe('jss-isolate', () => {
           }
         }
       })
-      setTimeout(done)
     })
 
     it('should not add selectors to the reset rule', () => {